OSX / WIndows compiler fixes

+++ src/genie/format/mpegg_p1/dataset/access_unit/block.cc
@@ -37,7 +37,7 @@ uint8_t Block::getDescID() const { return descriptor_ID; }
 
 // ---------------------------------------------------------------------------------------------------------------------
 
-uint32_t Block::getPayloadSize() const { return block_payload.size(); }
+uint32_t Block::getPayloadSize() const { return static_cast<uint32_t>(block_payload.size()); }
 
 // ---------------------------------------------------------------------------------------------------------------------
